requirements.txt
streamlit==1.32.2
openai==1.14.2
langchain==0.1.13
faiss-cpu==1.7.4
PyMuPDF==1.23.21
python-docx==1.1.0
python-pptx==0.6.23
